requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==23.3.0
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==4.1.5
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.16.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
uritemplate==4.1.1
